steps towards having ledger corss network
