+++ src/components/common/TmSessionHardware.vue
@@ -66,9 +66,11 @@
 
 <script>
 import TmBtn from "common/TmBtn"
-import { mapState } from "vuex"
+import { mapState, mapGetters } from "vuex"
 import HardwareState from "common/TmHardwareState"
 import SessionFrame from "common/SessionFrame"
+import { getAddressFromLedger } from "scripts/ledger"
+
 export default {
   name: `session-hardware`,
   components: {
@@ -86,6 +88,7 @@ export default {
   }),
   computed: {
     ...mapState([`session`]),
+    ...mapGetters({ networkId: `network` }),
     submitCaption() {
       return {
         connect: "Sign In",
@@ -109,7 +112,7 @@ export default {
       this.status = `detect`
       this.address = null
       try {
-        this.address = await this.$store.dispatch(`connectLedgerApp`)
+        this.address = await getAddressFromLedger(this.networkId)
         this.$router.push(`/`)
       } catch ({ message }) {
         this.status = `connect`
