+++ tests/unit/specs/components/ActionModal/components/ActionModal.spec.js
@@ -6,7 +6,7 @@ import { focusParentLast } from "src/directives"
 const localVue = createLocalVue()
 localVue.use(Vuelidate)
 localVue.directive("focus-last", focusParentLast)
-localVue.directive("focus", () => { })
+localVue.directive("focus", () => {})
 
 let mockSimulate = jest.fn(() => 123456)
 let mockSend = jest.fn(() => ({
@@ -138,7 +138,7 @@ describe(`ActionModal`, () => {
     const self = {
       $store,
       actionManager: {
-        setContext: () => { },
+        setContext: () => {},
         simulate: () => 12345,
         send: ActionManagerSend,
         simulateTxAPI: jest.fn(),
@@ -154,7 +154,7 @@ describe(`ActionModal`, () => {
       },
       submissionErrorPrefix: `PREFIX`,
       trackEvent: jest.fn(),
-      connectLedger: () => { },
+      connectLedger: () => {},
       onSendingFailed: jest.fn(),
       createContext: jest.fn()
     }
@@ -217,13 +217,6 @@ describe(`ActionModal`, () => {
     wrapper.vm.step = `sign`
     expect(wrapper.vm.selectedSignMethod).toBe(`ledger`)
     expect(wrapper.find(`#password`).exists()).toBe(false)
-  })
-
-  it(`should dispatch connectLedgerApp`, () => {
-    const $store = { dispatch: jest.fn() }
-    const self = { $store }
-    ActionModal.methods.connectLedger.call(self)
-    expect($store.dispatch).toHaveBeenCalledWith(`connectLedgerApp`)
   })
 
   describe(`should show the action modal`, () => {
@@ -611,7 +604,7 @@ describe(`ActionModal`, () => {
         isValidInput: jest.fn(() => true),
         selectedSignMethod: `local`,
         step: `details`,
-        validateChangeStep: jest.fn(() => { })
+        validateChangeStep: jest.fn(() => {})
       }
     })
